optimizations and fixes

+++ src/rsmod/flag/BlockAccessFlag.ts
@@ -1,7 +1,7 @@
 @final
 export default class BlockAccessFlag {
-    @inline static readonly BLOCK_NORTH: i32 = 0x1;
-    @inline static readonly BLOCK_EAST: i32 = 0x2;
-    @inline static readonly BLOCK_SOUTH: i32 = 0x4;
-    @inline static readonly BLOCK_WEST: i32 = 0x8;
+    @inline static readonly BLOCK_NORTH: i8 = 0x1;
+    @inline static readonly BLOCK_EAST: i8 = 0x2;
+    @inline static readonly BLOCK_SOUTH: i8 = 0x4;
+    @inline static readonly BLOCK_WEST: i8 = 0x8;
 }
